Ensure that structs have fields populated before we make arrays of them
Debugging an issue with the LOAD_FAST operation revealed that
  f->f_fastlocals
apparent wasn't being setup, and hence locals couldn't be read.

Debugging indicated that in PyEval_EvalFrameEx that it *was* set up
correctly, but not being seen by the JIT-generated code, with a big
discrepancy between the sizes of PyFrameObject:
in libpython:
  (gdb) p f
  $42 = (PyFrameObject *) 0x8554c0
  (gdb) p sizeof(*f)
  $43 = 384
whereas in the JITted code:
  (gdb) down
  #0  my_eval (f=0x8554c0, throwflag=<optimized out>) at /tmp/fake-eval.c:452
  452	  if (Py_EnterRecursiveCall ("")) goto Py_EnterRecursiveCall_failed; else goto Py_EnterRecursiveCall_succeeded;
  (gdb) p f
  $44 = (struct PyFrameObject *) 0x8554c0
  (gdb) p sizeof(*f)
  $45 = 144
which is (384 - 144) = 240 bytes smaller, which is the size of the
  PyTryBlock f_blockstack[CO_MAXBLOCKS];
field, and that the address of f_localsplus didn't agree between libpython
  (gdb) p &f->f_localsplus
  $56 = (PyObject *(*)[1]) 0x855638
  (gdb) down
  #0  my_eval (f=0x8554c0, throwflag=<optimized out>) at /tmp/fake-eval.c:452
  452	  if (Py_EnterRecursiveCall ("")) goto Py_EnterRecursiveCall_failed; else goto Py_EnterRecursiveCall_succeeded;
  (gdb) p &f->f_localsplus
  $57 = (struct PyObject *(*)[1]) 0x855548
and the JIT-generated code; when supposedly reading from f_localsplus it
was actually reading from the f_blockstack array.

Stepping through type-creation in libgccjit revealed that when
layout_type was called on the PyTryBlock array it saw a TYPE_SIZE of 0 for
the struct: the array type was being created *before* the fields of
PyTryBlock were being set up and hence has a bogus size.

The simplest solution is to require structs to have their fields set
before arrays of them can be created, which is now required as of
libgccjit commit 52023e93b077fc449eb52e3a8356ac2deab6dd4f.

This requires us to reorganize how we create types in coconut: given
that creation of types and population of fields is intermingled in
coconut.ir.compiler.Types, we now record a trace of events on the
coconut.ir.IrTypes object, consisting of type creations and field
population, and play this back in coconut.backend.libgccjit.GccJitBackend,
thus ensuring a valid ordering.

With this, f_localsplus can be accessed correctly from JITted code.